+++ lib/responsive_media.dart
@@ -1,2 +1,3 @@
 export 'src/responsive_media.dart';
 export 'src/extensions.dart';
+export 'advanced_breakpoints.dart';